improve the ITG3200 structure
