+++ AeroQuad/AeroQuad.ino
@@ -40,8 +40,8 @@
 
 // Mega platform
 //#define AeroQuadMega_v1     // Arduino Mega with AeroQuad Shield v1.7 and below
-//#define AeroQuadMega_v2     // Arduino Mega with AeroQuad Shield v2.0
-#define AeroQuadMega_v21    // Arduino Mega with AeroQuad Shield v2.1
+#define AeroQuadMega_v2     // Arduino Mega with AeroQuad Shield v2.0
+//#define AeroQuadMega_v21    // Arduino Mega with AeroQuad Shield v2.1
 //#define AeroQuadMega_Wii    // Arduino Mega with Wii Sensors and AeroQuad Shield v2.x
 //#define ArduCopter          // ArduPilot Mega (APM) with Oilpan Sensor Board
 //#define AeroQuadMega_CHR6DM // Clean Arduino Mega with CHR6DM as IMU/heading ref.
@@ -52,13 +52,13 @@
  *********************** Define Flight Configuration ************************
  ****************************************************************************/
 // Use only one of the following definitions
-//#define quadXConfig
+#define quadXConfig
 //#define quadPlusConfig
 //#define hexPlusConfig
 //#define hexXConfig      // EXPERIMENTAL: not completely re-tested
 //#define triConfig
 //#define quadY4Config
-#define hexY6Config
+//#define hexY6Config
 //#define octoX8Config
 //#define octoPlusConfig  // EXPERIMENTAL: not completely re-tested
 //#define octoXConfig     // EXPERIMENTAL: not completely re-tested
@@ -90,7 +90,7 @@
 // *******************************************************************************************************************************
 //#define HeadingMagHold // Enables Magnetometer, gets automatically selected if CHR6DM is defined
 #define AltitudeHoldBaro // Enables BMP085 Barometer (experimental, use at your own risk)
-//#define AltitudeHoldRangeFinder // EXPERIMENTAL : Enable altitude hold with range finder
+#define AltitudeHoldRangeFinder // EXPERIMENTAL : Enable altitude hold with range finder
 //#define RateModeOnly // Use this if you only have a gyro sensor, this will disable any attitude modes.
 
 //
@@ -352,6 +352,7 @@
   #include <Device_I2C.h>
 
   // Gyroscope declaration
+  #define ITG3200_ADDRESS_ALTERNATE
   #include <Gyroscope_ITG3200.h>
 
   // Accelerometer declaration
@@ -384,7 +385,6 @@
    * Put AeroQuad_Mini specific intialization need here
    */
   void initPlatform() {
-    gyroAddress = ITG3200_ADDRESS-1;
 
     pinMode(LED_Red, OUTPUT);
     digitalWrite(LED_Red, LOW);
@@ -542,6 +542,7 @@
   #include <Device_I2C.h>
 
   // Gyroscope declaration
+  #define ITG3200_ADDRESS_ALTERNATE
   #include <Gyroscope_ITG3200_9DOF.h>
 
   // Accelerometer declaration
