+++ Libraries/AQ_Gyroscope/Gyroscope_ITG3200Common.h
@@ -1,0 +1,116 @@
+/*
+  AeroQuad v3.0 - May 2011
+  www.AeroQuad.com
+  Copyright (c) 2011 Ted Carancho.  All rights reserved.
+  An Open Source Arduino based multicopter.
+ 
+  This program is free software: you can redistribute it and/or modify 
+  it under the terms of the GNU General Public License as published by 
+  the Free Software Foundation, either version 3 of the License, or 
+  (at your option) any later version. 
+
+  This program is distributed in the hope that it will be useful, 
+  but WITHOUT ANY WARRANTY; without even the implied warranty of 
+  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the 
+  GNU General Public License for more details. 
+
+  You should have received a copy of the GNU General Public License 
+  along with this program. If not, see <http://www.gnu.org/licenses/>. 
+*/
+
+#ifndef _AEROQUAD_GYROSCOPE_ITG3200_COMMON_H_
+#define _AEROQUAD_GYROSCOPE_ITG3200_COMMON_H_
+
+#include <Gyroscope.h>
+
+#ifdef ITG3200_ADDRESS_ALTERNATE
+  #define ITG3200_ADDRESS					0x68
+#else
+  #define ITG3200_ADDRESS					0x69
+#endif
+
+#define ITG3200_IDENTITY 0x69
+#define ITG3200_MEMORY_ADDRESS			0x1D
+#define ITG3200_BUFFER_SIZE				6
+#define ITG3200_RESET_ADDRESS			0x3E
+#define ITG3200_RESET_VALUE				0x80
+#define ITG3200_LOW_PASS_FILTER_ADDR	0x16
+#define ITG3200_LOW_PASS_FILTER_VALUE	0x1D	// 10Hz low pass filter
+#define ITG3200_OSCILLATOR_ADDR			0x3E
+#define ITG3200_OSCILLATOR_VALUE		0x01	// use X gyro oscillator
+#define ITG3200_SCALE_TO_RADIANS		823.626831 // 14.375 LSBs per °/sec, / Pi / 180
+
+
+
+
+void measureSpecificGyroADC(int *gyroADC);
+void measureSpecificGyroSum();
+void evaluateSpecificGyroRate(int *gyroADC);
+
+void initializeGyro() {
+
+  if (readWhoI2C(ITG3200_ADDRESS) == ITG3200_IDENTITY) {
+	vehicleState |= GYRO_DETECTED;
+  }
+	
+  gyroScaleFactor = radians(1.0 / 14.375);  //  ITG3200 14.375 LSBs per °/sec
+  updateRegisterI2C(ITG3200_ADDRESS, ITG3200_RESET_ADDRESS, ITG3200_RESET_VALUE); // send a reset to the device
+  updateRegisterI2C(ITG3200_ADDRESS, ITG3200_LOW_PASS_FILTER_ADDR, ITG3200_LOW_PASS_FILTER_VALUE); // 10Hz low pass filter
+  updateRegisterI2C(ITG3200_ADDRESS, ITG3200_RESET_ADDRESS, ITG3200_OSCILLATOR_VALUE); // use internal oscillator 
+}
+
+
+void measureGyro() {
+
+  sendByteI2C(ITG3200_ADDRESS, ITG3200_MEMORY_ADDRESS);
+  Wire.requestFrom(ITG3200_ADDRESS, ITG3200_BUFFER_SIZE);
+
+  int gyroADC[3];
+  measureSpecificGyroADC(gyroADC);
+
+  for (byte axis = 0; axis <= ZAXIS; axis++) {
+    gyroRate[axis] = filterSmooth(gyroADC[axis] * gyroScaleFactor, gyroRate[axis], gyroSmoothFactor);
+  }
+ 
+  // Measure gyro heading
+  long int currentTime = micros();
+  if (gyroRate[ZAXIS] > radians(1.0) || gyroRate[ZAXIS] < radians(-1.0)) {
+    gyroHeading += gyroRate[ZAXIS] * ((currentTime - gyroLastMesuredTime) / 1000000.0);
+  }
+  gyroLastMesuredTime = currentTime;
+}
+
+void measureGyroSum() {
+
+  sendByteI2C(ITG3200_ADDRESS, ITG3200_MEMORY_ADDRESS);
+  Wire.requestFrom(ITG3200_ADDRESS, ITG3200_BUFFER_SIZE);
+  
+  measureSpecificGyroSum();
+  
+  gyroSampleCount++;
+}
+
+void evaluateGyroRate() {
+
+  int gyroADC[3];
+  evaluateSpecificGyroRate(gyroADC);
+  gyroSample[XAXIS] = 0;
+  gyroSample[YAXIS] = 0;
+  gyroSample[ZAXIS] = 0;
+  gyroSampleCount = 0;
+
+  for (byte axis = 0; axis <= ZAXIS; axis++) {
+    gyroRate[axis] = filterSmooth(gyroADC[axis] * gyroScaleFactor, gyroRate[axis], gyroSmoothFactor);
+  }
+  
+  // Measure gyro heading
+  long int currentTime = micros();
+  if (gyroRate[ZAXIS] > radians(1.0) || gyroRate[ZAXIS] < radians(-1.0)) {
+    gyroHeading += gyroRate[ZAXIS] * ((currentTime - gyroLastMesuredTime) / 1000000.0);
+  }
+  gyroLastMesuredTime = currentTime;
+}
+
+
+
+#endif
