--- Libraries/AQ_Gyroscope/Gyroscope_ITG3200Defines.h
@@ -1,34 +1,0 @@
-/*
-  AeroQuad v3.0 - May 2011
-  www.AeroQuad.com
-  Copyright (c) 2011 Ted Carancho.  All rights reserved.
-  An Open Source Arduino based multicopter.
- 
-  This program is free software: you can redistribute it and/or modify 
-  it under the terms of the GNU General Public License as published by 
-  the Free Software Foundation, either version 3 of the License, or 
-  (at your option) any later version. 
-
-  This program is distributed in the hope that it will be useful, 
-  but WITHOUT ANY WARRANTY; without even the implied warranty of 
-  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the 
-  GNU General Public License for more details. 
-
-  You should have received a copy of the GNU General Public License 
-  along with this program. If not, see <http://www.gnu.org/licenses/>. 
-*/
-
-#ifndef _AEROQUAD_GYROSCOPE_ITG3200_DEFINES_H_
-#define _AEROQUAD_GYROSCOPE_ITG3200_DEFINES_H_
-
-#define ITG3200_MEMORY_ADDRESS			0x1D
-#define ITG3200_BUFFER_SIZE				6
-#define ITG3200_RESET_ADDRESS			0x3E
-#define ITG3200_RESET_VALUE				0x80
-#define ITG3200_LOW_PASS_FILTER_ADDR	0x16
-#define ITG3200_LOW_PASS_FILTER_VALUE	0x1D	// 10Hz low pass filter
-#define ITG3200_OSCILLATOR_ADDR			0x3E
-#define ITG3200_OSCILLATOR_VALUE		0x01	// use X gyro oscillator
-#define ITG3200_SCALE_TO_RADIANS		823.626831 // 14.375 LSBs per °/sec, / Pi / 180
-
-#endif
